api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
psycopg[binary]==3.2.4
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
pgvector==0.3.6
pymupdf4llm==0.0.17
docling==2.14.0
openai==1.57.2
httpx==0.28.0
python-multipart==0.0.20
tiktoken==0.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langgraph==1.1.0
langchain-core==0.3.45
langchain-openai==0.3.9
langgraph-checkpoint-postgres==2.0.6
